+++ src/plugin.h
@@ -5,5 +5,10 @@
 
 #include "LoBBSModule.h"
 
-#pragma MPM_MODULE(LoBBSModule, lobbsPlugin)
+/**
+ * This plugin registers a Module with Meshtastic. Modules are the primary way
+ * to extend Meshtastic with new functionality. Modules are registered with the
+ * MPM_REGISTER_MESHTASTIC_MODULE comment directive below.
+ */
+// MPM_REGISTER_MESHTASTIC_MODULE: LoBBSModule, lobbsModule, []